Make devtools operate without chroot directories present
To be able to program images when servo is controlled by the
Beaglebone board, the cros-devutils suite must be capable of operating
not only ouside chroot, but also in a situation where there is no
Chrome OS chroot directory tree present.

This change prevents the exception previously thrown in this case.

BUG=chrome-os-partner:15610
TEST=manual

   . programming DUT from inside chroot still works (the following
     command succeeds):

    $ cros_write_firmware -b daisy -w usb -d exynos5250-snow -F spi -i \
          /build/daisy/firmware/image-snow.bin -V

   . unittests still pass

    $ FEATURES=test sudo -E emerge cros-devutils

   . programming when servod is hosted by Beaglebone also works

Change-Id: I7a075b92898f93b1447bed70eca27a6510329482
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43499
Reviewed-by: Simon Glass <sjg@chromium.org>