Make sure cros_write_firmware reports actual status
The cros_write_firmware/cros_bundle_firmware utilities (both using
write_firmware.py) when programming ARM targets make the best effort
but do not have actual information about the flash programming
success/failure.

A recent enhancement of the servo driver allows the client to access
the CPU console output, where the programming success/failure is
reported. This change makes use of that enhancement.

When an ARM bootprom is being programmed and the -V (verify) option is
included, the programmer starts console output capture before
programming and examines the console stream generated while
programming, looking for the strings confirming success of failure of
the procedure.

Some refactoring is also thrown in:

 . the update and verify flags are stored in the Write object instead
   of being passed through multiple layers

 . _PrepareFlasher() is a private method

 . DutControl() becomes a public method

BUG=chrome-os-partner:15610
TEST=manual
 . run the following commands:
   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow -F spi -i \
           /build/daisy/firmware/image-snow.bin -V

   - succeeds

   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow -F spi -i \
          /build/daisy/firmware/image-snow.bin -V \
          -U /build/daisy/firmware/ec.bin

   - fails with "Target console dead, programming failed" error
     message (notice the bogus flasher)

Change-Id: I08c3d26833e842d23608d9c1b78000ff9cf32ae3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43160
Reviewed-by: Simon Glass <sjg@chromium.org>