dev_install: dereference symlinks
If the files in /etc/make.profile are symlinks, dereference them.
Otherwise, when we try to update/modify those files, we get an
error because they're pointing to files in /.

This is needed now that dev-install installs into /usr/share and
then symlinks its files into /etc.

BUG=chromium-os:38789
TEST=`dev_install` on board now works (no longer errors out on read-only rootfs)

Change-Id: Ic52e635c08cd5d33ada7c75377ed5bb2fada8c44
Reviewed-on: https://gerrit.chromium.org/gerrit/43189
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>