Make test failures trigger ebuild failures.
Many crosdev tests use the doctest.testmod() method to run inlined
tests. The problem is that test failures go unnoticed, as the return
value of the testmod() method is ignored. It is possible to invoke
this method such that it would throw an exception on any test failure,
but this would prevent the following tests from running.

A better solution is to assert in number of encountered failures is
not zero.

BUG=none
TEST=manual
    . try running

     $ FEATURES=test sudo -E emerge cros-devutils

    twice, first time with the code as submitted(it passes) and
    another time with one of asserts replaced with

    'assert doctest.testmod().failed != 0'

    in which case the ebuild reports failure.

Change-Id: I379b0f6f6f9286f4e65cc79dcef17c745b50b25b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43365
Reviewed-by: Simon Glass <sjg@chromium.org>