Avoid leaving temp directory behind unintentionally
In some cases when cros utilities are interrupted, the temporary
directory is left behind in the /tmp folder. It is not completely
clear why it happens, as the directory is supposed to be removed by
the __exit__() method of the Tools class.

This patch streamlines the logic (if user provides the temp directory
name, it will not be deleted even if -p flag is not given), and makes
sure that the directory uses a distinct name, so is there are some
leftovers it is easy to associate them with this utility.

BUG=None
TEST=manual

   . made sure that

    $ FEATURES=test sudo -E emerge cros-devutils

   still succeeds

   . tried different variants of invoking cros_write_firmware

   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow \
       -F spi -i /build/daisy/firmware/image-snow.bin -V

    no tmp directory left behind

   -----
   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow \
       -F spi -i /build/daisy/firmware/image-snow.bin -V

    /tmp/cros-dev.SicHJJ left behind

   -----
   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow \
       -F spi -i /build/daisy/firmware/image-snow.bin -V -O /tmp/xyz

   /tmp/xyz was created and left behind
   -----
   $ cros_write_firmware -b daisy -w usb -d exynos5250-snow \
       -F spi -i /build/daisy/firmware/image-snow.bin -V -O /xyz/zyx
   fails with error message:

   Cannot make output directory '/xyz/zyx': 'Permission denied'

Change-Id: Ia365fdb8eea8e994ea3c9453c61ca5ee51ac6e3c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43428