cros_bundle_firmware: Remove FDT when extracting U-Boot from image
To flash an image onto a board we download a flasher to the board, complete
with a script telling it what to do. The flasher is itself just U-Boot, and
we doctor up the FDT with a suitable script.

If no u-boot.bin is available, it is convenient to be able to extract the
U-Boot from the image being flashed. It is exactly the same binary, so there
is no problem with doing this.

However, at present the code that does this doesn't correctly determine the
end of U-Boot, partly since there is no defined way to do this. The intent
is that we will start using the fdtmap instead of FMAP, and so will have
access to that information. Until then, we need a workaround.

Search for the FDT magic word at the end of the extract image portion. Since
the FDT is attached to the end of U-Boot, this magic word marks the end of
the original u-boot.bin which was placed in the image. Use this smaller
portion as our u-boot.bin. When the flasher code adds its specially
doctored-up FDT, it will go in right place, and the flasher will work.

BUG=TBD, Vadim?
TEST=manual
Without this change:

$ cros_write_firmware -b daisy -w usb \
	-d /build/daisy/firmware/dts/exynos5250-snow.dts \
	-F spi -i /build/daisy/firmware/image-snow.bin -V -D -O out

just does a verified boot into the kernel.

With this change, it runs the flasher.

Change-Id: I53a50fd9bb19ec2ff2dc2e88b0dc894937678957
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43503