builder.py: use anchored globbing to match DEFAULT_INSTALL_MASK
We exclude files according to DEFAULT_INSTALL_MASK for binpkgs that
gmerge generates.  Using unanchored globbing for the exclusion results
in unwanted exclusion; for example, the entry "/firmware" will exclude
both "/firmware" and "/lib/firmware", the latter of which we want to
keep, as its prefix is not wildcarded.

This change adds the "--anchored" flag to the tar command to make sure
we match a prefix of the file's path.

BUG=chromium-os:21099 (firmware packaging issues)
TEST=local build, run on host machine

Change-Id: Iaf2fc8092ebd7a1bbaa6330d0d6035a651c1215f
Reviewed-on: https://gerrit.chromium.org/gerrit/44614
Commit-Queue: John Sheu <sheu@chromium.org>
Reviewed-by: John Sheu <sheu@chromium.org>
Tested-by: John Sheu <sheu@chromium.org>