cros_bundle_firmware: Automatically select BL1/BL2 for daisy/peach
Rather than using the wrong default, select the Exynos model name based on
the board type.

BUG=chrome-os-partner:18451
TEST=manual
Run cros_bundle_firmware with '-b peach_pit' and see that it no longer
complains about missing Exynos 5250 files. Run with '-b daisy' and see
that it picks up the 5250 files.

Force 'model' to none, and see that when no '--bl1' is provided, we get
a fairly sensible error:

    Cannot find value for entry property 'exynos-bl1'

Change-Id: I3e83fe90f5ab509dc3a5a3bfd916529f457f57a9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49398
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>