cros_bundle_firmware: Add 'align' property
While the size property helps avoid the pain of specifying the start
of each region, it does not completely solve the problem, since some
regions must be aligned to a particular boundary.

Add an 'align' property to deal with this problem. The requested
alignment of a region will be respected when using the 'size'
property. When using 'reg', where both the offset and size of the
region are given explicitly, the alignment will be checked and
cros_bundle_firmware will produce an error if it does not match.

BUG=none
TEST=manual
Add 'align = <0x8000>' in a section and see that a gap is created
before it.

P 00000000  00002000   BL1 PRE_BOOT
! 00002000  00006000   <gap>
P 00008000  00004000   BL2 SPL

With 'align = <0x4000>' the gap closes a bit:

P 00000000  00002000   BL1 PRE_BOOT
! 00002000  00002000   <gap>
P 00004000  00004000   BL2 SPL

With 'align = <0x2000>' (and also with no align property) the gap
closes:

P 00000000  00002000   BL1 PRE_BOOT
P 00002000  00004000   BL2 SPL

Change-Id: I8463278cb75e329743a67702b99bdc432c1d46a8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49400
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>