cros_bundle_firmware: Tidy up servo messages
When using '--servo none' we should not say that we are using servo to
reset the board.

BUG=chromium:216858
TEST=manual
Try on link with '--servo none' and check that the reset message no longer
appears, and link is not reset. Try with '--servo 7777' and see that it
does print the message and reset.

Also try the same with snow.

Change-Id: I9d28ed69e6cbe5f5bf169a0921b3f45d6277c1f6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49402