drop: cros_bundle_firmware: Pass device tree files through cpp
The kernel now passes device tree files through the C preprocessor
to allow #include and #define. Some U-Boot files are starting to do
the same, and also it would be nice to be able to use the same device
tree file for Chrome OS and upstream. We can achieve the latter with
something like this in the .dts file:

	#ifdef CONFIG_CHROMEOS
	/include/ "flashmap-exynos-ro.dtsi"
	/include/ "flashmap-4mb-rw.dtsi"
	/include/ "chromeos-exynos.dtsi"
	#endif

We could perhaps add an option to enable CONFIG_CHROMEOS, but for now
assume it is always set.

The -Ulinux is to deal with a gcc feature that it pre-defines 'linux'
and we don't want that word changed in our device tree files.

Note: We want to stop using this flow from ebuilds but this gets us
running for now. See crbug.com/246291

BUG=chromium:246291
TEST=manual
Add the above fragment to a device tree file. See that it is correctly
compiled by dtc, and the cpp has clearly been run on the source file.

Change-Id: I283872108b7181e475b69eb8ab67e72a294fdf10
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56115
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>