cros_bundle_firmware: Avoid hard-coding the U-Boot entry point
We want to find the cold boot entry point for U-Boot x86. This is not
necessarily at the same address. So instead of using a hard-coded position,
search for the first instruction of the code boot region.

This is really just a hack, since there is no guarantee that U-Boot won't
change again the future, although it is very stable.

Still, the existing code is even worse, since it uses a hard-coded offset!

BUG=chrome-os-partner:19534
BUG=chromium:245311
TEST=manual
Build and boot on link. Type 'vboot_twostop'. See that it correctly identifies
itself as RO firmware.

Change-Id: Id586ab319d2b6d49decce7caab5e7c2efbb5fc29
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57050
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>