paycheck/applier: stop using fallocate
It appears that fallocate is not supported on certain filesystems (e.g.
failing on paygen genati instances). It also appears that we can do just
fine without it when applying a full update, without significant impact
on performance.  In the interest of making the update_payload library
more portable, we therefore drop the fallocate call and replace it with
a simple create/truncate code (all Python standard lib calls).

BUG=chromium:244091
TEST=Integration tests

Change-Id: I57f7dec19e8131050c1d56d891e486714d4f7128
Reviewed-on: https://gerrit.chromium.org/gerrit/56762
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Matt Tennant <mtennant@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>