Adjust the timing of log rotation.
The intent for log rotation was "midnight between Friday and
Saturday", so that if an event of interest occurred over the
weekend, it would (likely) be contained entirely in the current log
on Monday morning.  Because of ambiguity in the Python
documentation, the actual time ended up being "midnight between
Saturday and Sunday".

This change adjusts the rotation timing to match the original intent.

BUG=chromium:219168
TEST=None

Change-Id: Ib82714380b1525f7630d509a2842002253726d34
Reviewed-on: https://gerrit.chromium.org/gerrit/57491
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>