cros_bunle_firmware: Separate Exynos bl2 processing into its own module
This change is a pure refactoring exercise - Exynos BL2 processing is
being moved into a separate module to make it easier to enhance it to
process different types of SPLs. Moving other Exynos specific
processing into this module will happen later.

The relevant functions are being moved to the new file, renamed for
clarity and become methods of the new class, ExynosBl2.

There are only three attributes in the new class, one of them (
self.spl_source) not used anywhere else in Bundle.

BUG=chromium:214441
TEST=manual

   . used a standard script to run cros_bundle_firmware to create a
     bootable image for peach_pit, verified that the image boots.

  . observed that unittests still succeed:
   $ FEATURES=test sudo -E emerge cros-devutils

Change-Id: Ie1a1bc43976f4c9b1f09c498e7dace42385aed35
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58029
Reviewed-by: Simon Glass <sjg@chromium.org>