cros_bundle_firmware: Permit sections to end unaligned
It turns out to be useful to have a section size not match the alignment
for its start position. There is no particular requirement here, so don't
check size alignment.

BUG=none
TEST=manual
Run with an updated flashmap where the 'fdtmap' section is aligned to
64KB but only 32KB in size. See that we no longer get an error in this
case.

An example is:

ro-fdtmap {
	label = "fdtmap";

	/* We encourage to align FMAP partition in as large
	* block as possible so that flashrom can find it soon.
	* For example, aligning to 512KB is better than to
	* 256KB. */

	size = <0x1000>;
	align = <0x10000>;
};

Previously this would give an error, but now it works.

Change-Id: If6ef9dddc06d0f262a329b6f0f44df6d48ed86ac
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58629