Derive Exynos component sizes from the device tree
The heuristics used for determining the u-boot image components
boundaries are failing when SPL is enhanced to provide serial console
output. The 0xea000014 signature is present in the SPL blob in that
case. It is not aligned at 4 byte boundary, but still trips the u-boot
base search code.

To make this a bit more robust, try using the device tree to find the
component sizes, if the appropriate nodes not found in the device tree
- fall back to the previously used heuristics and hope for the best.

BRANCH=
BUG=None
TEST=manual
  . make sure unit tests still pass
    $ FEATURES=test sudo -E emerge cros-devutils
  . generate an SD card image and boot it on pit
  . generate an SD flasher, boot it on pit, observe it flash new firmware
  . generate an SD card image and boot it on Snow

Change-Id: I937d76834cce62fcd5ed050726a010964d0fe94f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61526
Reviewed-by: Simon Glass <sjg@chromium.org>