cros_bundle_firmware: configure exynos flasher BL2
When creating a flasher image, ExynosFlashImage is passed the SPL binary
directly built with U-Boot to use as BL2, which has not had its machine
parameters configured.  By using this BL2 directly, the default values
in the machine parameters will be used, rather than the appropriate
configured values.

The default memory manufacturer for daisy is not correct for daisy_spring,
resulting in a flasher image that hangs in the memory initialization in
SPL.

When passed BL2, run it through the Exynos BL2 configuration to ensure the
machine parameters are updated.

BUG=none
BRANCH=none
TEST=`crosfw --board daisy_spring -VF` successfully flashes spring
`crosfw --board daisy -VF` successfully boots flasher on snow

Change-Id: I815b17aecbcc56fb62b462e0cb009d6e084fdcb8
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61681
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>