cros_bundle_firmware: Split RO and RW sections on eMMC
On the eMMC, if the flash map contains a wp-ro node to describe the size
of the read-only section, build a flasher that will flash the read-only
section to boot partition 1 and the read-write section to boot partition
2.

BUG=chromium:254233
BRANCH=none
TEST=On spring, cros_bundle_firmware -w usb -F sdmmc successfully writes
the RO section to boot partition 1, and the RW section to boot partition 2.
If /flash/wp-ro is removed from the flash map, the entire image is written
to boot partition 1, with a warning about being unable to detect the RO
section size.
SPI flashing is unaffected.

Change-Id: I061fb9b3c68b190c11dda2b09286453167c47681
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62665
Reviewed-by: Simon Glass <sjg@chromium.org>