cros_bundle_firmware: Provide the updated FDT to write_firmware
In some cases write_firmware needs to see the updated FDT, not the original
one that was provided by the user. An example of this is if write_firmware
wants to look at some information added by the pack_firmware. An example of
useful information would be the 'used' property in a later commit, where
write_firmware wants to write to the target only the portion of a region that
is actually used.

Since we only ever add things to the FDT, it should be harmless to share the
same FDT file.

The flasher makes a copy before it changes anything, so changes made by the
flasher will not affect the FDT that is returned from bundle_firmware.

BUG=chrome-os-partner:21115
TEST=manual
Tested with later commits.

Change-Id: I6e1f91653fb58dde8aaaa5c0a68b2b8ef9befd29
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62582