cros_bundle_firmware: Remove old RW device trees
We were planning on provding RW firmware with a slightly different device
tree than RO. I think this was so that we could avoid the coreboot hack of
jumping to a different image start to detect at run-time which firmware is
running.

However, it just adds complication and the resulting FDTs are currently not
used by any flashmap, so let's drop it.

BUG=none
TEST=FEATURES=test sudo -E  emerge cros-devutils
grep for 'updated-rw' in U-Boot source tree finds nothing

Change-Id: I17b4eb0c77961628c608d063eaa149f751cd0fb5
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62738
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>