cros_bundle_firmware: Allow a different Exynos SPL for each flashmap section
At present we expect to have only one SPL in an image - it is used to load
the read-only U-Boot.

With early-firmware-selection we are going to need to support an SPL for each
of the RW U-Boots also, since that is where SDRAM initialization will happen.

Adjust the code to remember the relationship between an SPL and the particular
flashmap section that created it, so we can have muiltiple SPLs and each can
be configured by the parameters in that flashmap section.

Also move the Exynos SPL/BL2 code into exynos.py, since this code belongs
there rather than in bundle_firmware.

The previously used GetBlobParams() is still present, so add a comment
clarifying that this can only get the parameters for a blob which is the same
every time it is used. For example, the U-Boot binary is inserted without
modification anywhere where the 'boot' blob is called for.

This prepares the way for adding hash support to SPL in a future commit.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils
'crosfw -b peach_pit -V' on pit and see that it still builds and boots.

Change-Id: I6d985f89a98c53a7579ab89773aaafabd8ec9ea6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62740