cros_bundle_firmware: Support early-firmware-selection blob types
When early-firmware-selection is enabled, we use the 'type,efs' field instead
of 'type'. This allows us to specify different blobs in that case.

Enable 'early-firmware-selection' as a boolean property in /chromeos-config
and provide a --efs option to force it to be enabled for convenience during
development.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils
Not specifically tested yet. Tested by later commits.

Change-Id: I7d77508d7c8cdc7fcd243b4240476a9c34231532
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62962