fdt: Add PutBool to write a boolean property
A boolean property is one that is True when present and False when absent.
Add a function to write a boolean property.

Using this definition it is not possible to change a property which is True
into one that is False, since fdtput does not support it. For now, we raise
an error since this functionality is not currently needed.

I did propose on LKML that a property with a single cell of value zero also
be considered False. There was not a lot of interest. We do that in U-Boot
but not in Linux.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils

Change-Id: I62430ef33db1e5070ff2c2e70506fddc97e339c4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62963