cros_bundle_firmware: Remove special case for default flashmap
Instead of passing the default flashmap as a special case to
PackFirmware.SelectFdt(), write the default flashmap to the FDT
directly. SelectFdt() will then do the right thing automatically.

One reason for this change is that we want to add other default
things to the device tree, not just the flashmap. For example, we
will soon implement values for the BL1 and BL2 addresses, and defaults
must be provided for booting upstream U-Boot.

Another reason is that we can remove all the special case code in
PackFirmware.SelectFdt().

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils
manual:

See that all of these create images which boot on the respective boards.
This tests Chrome OS and upstream version of each board we support.

$ crosfw -b peach_pit -V
$ crosfw -b daisy
$ crosfw -b daisy -V
$ crosfw -b link
$ crosfw -b link -V

Change-Id: I942afd7972f9da098445eeaf0f78de127464ed5a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62965