write_firmware: Correct handling of writing upstream U-Boot
The except clause did not cope completely with a missing property since the
error received when the flash/ node is present but the size property is not
is:

   KeyError: 'size'

and not a CmdError. This change is needed by the next commit which adds
properties to the /flash node.

BUG=none
TEST=manual

With upstream U-Boot:

$ crosfw -b daisy

See that it uploads to snow correctly.

Change-Id: I58bad1bc9d4f6c142fd8342216a292587b81a600
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62961