cros_bundle_firmware: Add a file column to the --map display
It is useful to see what files are packed into what sections of the flash.
Enhance -m to support this.

Note:
In most cases there is only a single file in each section - the only
exception at present is that the device tree is packed in with U-Boot
since we currently have no way to read this separately in SPL. If we adjust
U-Boot to read the device tree from the FDTMAP section then we could
potentially remove support for multiple images in a section which would
simplify many parts of the code.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils

Run cros_bundle_firmware with -m and see that filenames are now shown.

Change-Id: I399a2b30b8b061a4d7e9769b684524aced6d133d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63043