Add --nodeps to cros_workon_make.
Currently, cros_workon_make --install calculates dependencies, in case other
related packages need to be installed. This is not necessary and this is
not the intention of cros_workon_make. Add --nodeps to remove this logic
and speed up cros_workon_make by 3 seconds.

BUG=chromium:267928
TEST=Confirm cros_workon_make is 3 seconds faster.

Change-Id: Ic15c2987f9fc7a084a8f445277cfea936ca80b47
Reviewed-on: https://gerrit.chromium.org/gerrit/64426
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>