cros_bundle_firmware: Add RW SPL parameters
We need to add new machine parameters to tell SPL whether to jump to RW SPL
or not. If so it also needs the address (in IRAM) and size of that RW SPL
so that it can find it.

All of this is available in the device tree, so pull it from there.

In RW SPL we force loading from SPI flash. We don't support loading from
USB or any other media, so this needs to be hard-coded. This allows us to
USB-boot a board with RO firmware and have it still perform a full boot
through RW - without this change we just get a hard hang here, waiting on
USB input which never comes.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils
Perform an EFS boot on Pit and do resume testing, like this.

First use 'suspend_stress_test -c10' to test that s2r works OK.

Then corrupt the RW SPL:

memtester -p 0x206c000 8192B

(the RW SPL is stored at 0x0206d800 and is 16KB long)

Try s2r again - it now hangs on resume. I tried this sequence twice.

Also verified with messages printed by RO U-Boot:

readonly.c: vboot_stash_rw_spl: Stashed RW SPL for resume path at 206d800, size 4000, free 534

Change-Id: Id271f39b2a51d5742dc0d974ef50aa6794c28be4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64738