cros_bundle_firmware: Don't round up U-Boot size
SPL is told how large U-Boot is so that it knows how much to loads from
the boot media. For SD and eMMC we need to load a specific number of blocks,
so the size is rounded up to the nearest block size. This rounding was
done in cros_bundle_firmware to minimize the amount of messing around
required in SPL.

However, rounding up the U-Boot size is not a good idea if it is to be
hashed. It makes the hash incorrect and SPL cannot verify it since it
does not know the correct size.

Remove the rounding, and leave it to SPL to round up the size for eMMC
and SD card if it wants to. Then the hash we calculate can be verified
by SPL without any problem.

This requires a corresponding change in U-Boot, to come.

BUG=chrome-os-partner:21115
TEST=FEATURES=test sudo -E  emerge cros-devutils
Check that RW SPL can now verify the U-Boot hash.

This SPL message appears during boot, indicating that the hash was
verified.

SPI fast...loaded to 23e00000, size 6eee0, verified, jump

Change-Id: Id08ee446209f22783c94bb8b6d0ae38c8a34291f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64740