Set default artifacts for download and wait_for_status calls
download and wait_for_status used to have hard-coded value of artifacts. This
behavior was changed in CL:
https://gerrit.chromium.org/gerrit/#/c/64310
This change breaks lab until this CL in autotest is checked in:
https://gerrit.chromium.org/gerrit/#/c/64706
To make the TOT of devserver work with autotest in prod, this CL is created as
a temporary fix. After CL 64706 is merged and pushed to prod, this CL can be
reverted.

BUG=chromium:269197
TEST=tested with following links:
http://dshi.mtv:8094/download?archive_url=gs://chromeos-image-archive/lumpy-release/R30-4408.0.0/
http://dshi.mtv:8094/wait_for_status?archive_url=gs://chromeos-image-archive/lumpy-release/R30-4408.0.0/

Change-Id: Id884e16f4f168a608bbc24d10131c7934e35879d
Reviewed-on: https://gerrit.chromium.org/gerrit/64884
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>